feat: refactor Pokja service and controller to use DTO, add SQL scripts for new tables and update CSV data format

+++ src/main/resources/db/changelog/sql/001-create-kategori_instansi.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS kategori_instansi (
-    id_kategori_instansi BIGINT PRIMARY KEY,
+    id_kategori_instansi SERIAL PRIMARY KEY,
     kategori_instansi VARCHAR(255)
 );
