+++ src/main/resources/db/changelog/sql/002-create-jenis_instansi.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS jenis_instansi (
-    id_jenis_instansi BIGINT PRIMARY KEY,
+    id_jenis_instansi SERIAL PRIMARY KEY,
     jenis_instansi VARCHAR(255)
 );
