+++ src/main/resources/db/changelog/sql/003-create-wilayah_kerja.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS wilayah_kerja (
-    id_wilayah BIGINT PRIMARY KEY,
+    id_wilayah SERIAL PRIMARY KEY,
     nama_wilayah VARCHAR(255)
 );
