+++ src/main/resources/db/changelog/sql/004-create-pokja.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS pokja (
-    id_pokja BIGINT PRIMARY KEY,
+    id_pokja SERIAL PRIMARY KEY,
     nama_pokja VARCHAR(255)
 );
