+++ src/main/resources/db/changelog/sql/005-create-instansi.sql
@@ -1,10 +1,10 @@
 CREATE TABLE IF NOT EXISTS instansi (
-    id_instansi BIGINT PRIMARY KEY,
+    id_instansi SERIAL PRIMARY KEY,
     nama_instansi VARCHAR(255),
-    id_kategori_instansi BIGINT,
-    id_jenis_instansi BIGINT,
-    id_wilayah_kerja BIGINT,
-    id_pokja BIGINT
+    id_kategori_instansi INT,
+    id_jenis_instansi INT,
+    id_wilayah_kerja INT,
+    id_pokja INT
 );
 ALTER TABLE instansi ADD CONSTRAINT fk__id_kategori_instansi FOREIGN KEY (id_kategori_instansi) REFERENCES kategori_instansi(id_kategori_instansi);
 ALTER TABLE instansi ADD CONSTRAINT fk__id_jenis_instansi FOREIGN KEY (id_jenis_instansi) REFERENCES jenis_instansi(id_jenis_instansi);
