+++ src/main/resources/db/changelog/sql/008-create-jenjang.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS jenjang (
-    id_jenjang BIGINT PRIMARY KEY,
+    id_jenjang SERIAL PRIMARY KEY,
     jenjang VARCHAR(255)
 );
